bookMapperXML

+++ src/main/java/com/bookStore/controller/AddressController.java
@@ -26,24 +26,20 @@ public class AddressController {
     @Autowired
     private AddressService addressService;
 
-    //方法参数说明，name参数名；value参数说明，备注；dataType参数类型；required 是否必传；defaultValue 默认值
-    @ApiImplicitParam(name = "address", value = "新增地址数据")
     //说明是什么方法(可以理解为方法注释)
     @ApiOperation(value = "添加地址", notes = "添加地址")
     @PostMapping("addAddress")
     public RestResult addAddress(@RequestBody @Valid Address address) {
         RestResult restResult = new RestResult(ResultCode.OPERATION_FAILURE);
-
         Long userId =  ThreadLocalUtil.get();
         int rows = addressService.addAddress(userId, address);
+        //插入成功
         if (rows > 0) {
             restResult = new RestResult(ResultCode.SUCCESS);
         }
         return restResult;
     }
 
-    //方法参数说明，name参数名；value参数说明，备注；dataType参数类型；required 是否必传；defaultValue 默认值
-    @ApiImplicitParam(name = "address", value = "修改的地址数据")
     //说明是什么方法(可以理解为方法注释)
     @ApiOperation(value = "修改地址", notes = "修改地址")
     @PutMapping("updateAddress")
@@ -61,14 +57,16 @@ public class AddressController {
         return restResult;
     }
 
-    //方法参数说明，name参数名；value参数说明，备注；dataType参数类型；required 是否必传；defaultValue 默认值
-    @ApiImplicitParam(name = "addressId", value = "修改地址的Id")
     //说明是什么方法(可以理解为方法注释)
     @ApiOperation(value = "删除地址", notes = "删除地址")
     @DeleteMapping("deleteAddress")
     public RestResult deleteAddress(@NotNull Long addressId) {
-        Long userId = (Long) ThreadLocalUtil.get();
         RestResult restResult = new RestResult(ResultCode.OPERATION_FAILURE);
+        //没有传入地址Id
+        if(addressId == null){
+            return restResult;
+        }
+        Long userId = (Long) ThreadLocalUtil.get();
         int rows = addressService.deleteAddress(userId, addressId);
         if (rows > 0) {
             restResult = new RestResult(ResultCode.SUCCESS);
